requirements.txt
flask==2.3.2
flask-cors==4.0.0
google-api-python-client==2.88.0
google-auth==2.19.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
openai==0.27.8
pytest==7.3.1
python-dotenv==1.0.0